feat: partial config reload resistant image cycling; refactor: start and stop loop triggers
The logic for image cycling is redone. It now relies on
"window-config-reloaded" to keep track of config reloads, both partial
and full. For full, current_image won't have a value so we start the
loop. For partial, we check if there is a loop existing and start one;
otherwise, we schedule a call after an interval. This avoids overlapping
loops and differentiates between partial and full reloads, auto cycling
and manual cycling.